feat: made viraliseai component stream as well
