+++ app/api/content/route.ts
@@ -1,7 +1,7 @@
 // Use getAuth for server-side
 import { auth } from "@clerk/nextjs/server";
 import { NextResponse } from "next/server";
-import OpenAIService from '@/app/core/openAIService';
+import OpenAIService from "@/app/core/openAIService";
 
 // Define the POST request handler for the API route
 export async function POST(req: Request) {
@@ -14,7 +14,18 @@ export async function POST(req: Request) {
     }
 
     const oAIService = OpenAIService.getInstance(process.env.OPENAI_API_KEY);
-    const stream = await oAIService.generateTextStream(messages);
+    const response = await oAIService.generateTextStream(messages);
+    const encoder = new TextEncoder();
+    const stream = new ReadableStream({
+      async start(controller) {
+        for await (const chunk of response) {
+          controller.enqueue(
+            encoder.encode(chunk.choices[0]?.delta?.content || "")
+          );
+        }
+        controller.close();
+      },
+    });
 
     return new NextResponse(stream, {
       headers: {
@@ -23,7 +34,6 @@ export async function POST(req: Request) {
         Connection: "keep-alive",
       },
     });
-
   } catch (error) {
     console.log("[CONVERSATION_ERROR]", error);
     return new NextResponse(`Internal error ${error}`, { status: 500 });
